refactor Universe
